requirements.txt
Flask
torch
ultralytics
opencv-python>=4.6
numpy
matplotlib
Pillow
tensorflow-cpu
Werkzeug
gunicorn